requirements.txt
# Energy Forecasting Module Dependencies

# ============================================================================
# MACHINE LEARNING
# ============================================================================
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0
scikit-learn>=1.3.0

# ============================================================================
# DATA PROCESSING
# ============================================================================
pandas>=2.0.0
numpy>=1.24.0

# ============================================================================
# HYPERPARAMETER OPTIMIZATION
# ============================================================================
optuna>=3.0.0

# ============================================================================
# CALENDAR / HOLIDAYS
# ============================================================================
holidays>=0.40

# ============================================================================
# UTILITIES
# ============================================================================
python-dotenv>=1.0.0
pytz>=2024.1
tqdm>=4.66.0
joblib>=1.3.0
